requirements.txt
matplotlib~=3.7.1
speedtest-cli~=2.1.3